issue-96: Orbital station description was added

+++ SpaceInvaders/src/main/java/com/gmail/yaroslavlancelot/spaceinvaders/objects/objects/loading/units/UnitLoader.java
@@ -40,7 +40,7 @@ public class UnitLoader {
     public Integer speed;
 
     @Element(name = "reload_time")
-    public Double reload_time;
+    public Float reload_time;
 
     @Element(name = "sound")
     public String sound;
